requirements.txt
PyYAML==5.1.1
numpy==1.16.4
tabulate==0.8.3
